+++ settings.js
@@ -41,6 +41,20 @@ export const registerSettings = function () {
 		type: Boolean,
 	});
 
+	game.settings.register(modulename, "drop-item-size", {
+		name: i18n("MonksLittleDetails.drop-item-size.name"),
+		hint: i18n("MonksLittleDetails.drop-item-size.hint"),
+		scope: "world",
+		config: true,
+		range: {
+			min: 0.1,
+			max: 1.5,
+			step: 0.05,
+		},
+		default: 1,
+		type: Number,
+	});
+
 	game.settings.register(modulename, "allow-door", {
 		name: i18n("MonksActiveTiles.allow-door.name"),
 		hint: i18n("MonksActiveTiles.allow-door.hint"),